allow camera plugin to be installed when android api < 21 fix https://github.com/flutter/flutter/issues/30854
