+++ packages/camera/example/android/app/src/main/AndroidManifest.xml
@@ -1,8 +1,9 @@
 <manifest xmlns:android="http://schemas.android.com/apk/res/android"
+    xmlns:tools="http://schemas.android.com/tools"
     package="io.flutter.plugins.cameraexample">
 
     <uses-permission android:name="android.permission.INTERNET"/>
-
+    <uses-sdk tools:overrideLibrary="io.flutter.plugins.camera"/>
     <uses-feature
         android:name="android.hardware.camera"
         android:required="true" />
